+++ cli/util/server.go
@@ -1,10 +1,10 @@
 package util
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 
 	"github.com/hasura/graphql-engine/cli/v2/internal/httpc"
@@ -98,16 +98,18 @@ func GetServerState(client *httpc.Client, endpoint string, hasMetadataV3 bool, l
 
 }
 
-func GetServerStatus(versionEndpoint string) (err error) {
-	resp, err := http.Get(versionEndpoint)
+func GetServerStatus(versionEndpoint string, httpClient *httpc.Client) (err error) {
+	req, err := http.NewRequest("GET", versionEndpoint, nil)
+	if err != nil {
+		return fmt.Errorf("failed to create GET request to %s: %w", versionEndpoint, err)
+	}
+	var responseBs bytes.Buffer
+	resp, err := httpClient.Do(context.Background(), req, &responseBs)
 	if err != nil {
 		return fmt.Errorf("making http request failed: %w", err)
 	}
-	defer resp.Body.Close()
-	// ignore error
-	body, _ := ioutil.ReadAll(resp.Body)
 	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("request failed: url: %s status code: %v status: %s \n%s", versionEndpoint, resp.StatusCode, resp.Status, string(body))
+		return fmt.Errorf("request failed: url: %s status code: %v status: %s \n%s", versionEndpoint, resp.StatusCode, resp.Status, responseBs.String())
 	}
 	return nil
 }